Added video to epub converter
This script is somewhat ridiculous. It takes a video and .vtt
subtitles file and converts it to an .epub file with images for each
subtitle. This makes it possible to "read" a TV episode on an E-Ink
tablet, for example.

I made this because I'm not good enough at listening to Japanese at
full speed. Now, I can "read the episode" first, then watch it later.